+++ CHANGES.md
@@ -1,4 +1,5 @@
 # 0.17
+- Add `runEnvironment` key to control environment variables
 
 # 0.16 2020-12-21
 - In plugin for intellij plugins: Added ability to instrument classes with not-null assertions and to compile .form files